Reorganized the terasort code into a separate file. Added a new utils.py file for common code. Used the dcos.metronome module for working with metronome jobs.
