+++ tests/jobs/json/hdfs-delete-terasort-files.json
@@ -8,6 +8,7 @@
     },
     "cpus": 0.1,
     "mem": 512,
+    "disk": 0,
     "user": "root",
     "restart": {
       "policy": "NEVER"
