+++ tests/test_spark.py
@@ -11,58 +11,45 @@ import dcos.config
 import dcos.http
 import dcos.package
 
-import dcosjob
 import logging
 import os
 import pytest
-import re
 import s3
 import shakedown
-import subprocess
-import time
 import urllib
 
-
-def _init_logging():
-    logging.basicConfig(level=logging.INFO)
-    logging.getLogger('dcos').setLevel(logging.WARNING)
-    logging.getLogger('requests').setLevel(logging.WARNING)
+import utils
+from utils import SPARK_PACKAGE_NAME
 
 
-_init_logging()
 LOGGER = logging.getLogger(__name__)
 THIS_DIR = os.path.dirname(os.path.abspath(__file__))
 DEFAULT_HDFS_TASK_COUNT=10
 HDFS_PACKAGE_NAME='beta-hdfs'
 HDFS_SERVICE_NAME='hdfs'
-SPARK_PACKAGE_NAME='spark'
-TERASORT_JAR='https://downloads.mesosphere.io/spark/examples/spark-terasort-1.0-jar-with-dependencies_2.11.jar'
-TERASORT_MAX_EXECUTOR_CORES=6
 
 
 def setup_module(module):
-    if _hdfs_enabled():
+    if utils.hdfs_enabled():
         _require_hdfs()
     _require_spark()
-    _require_spark_cli()
 
 
 def teardown_module(module):
-    if _do_teardown():
-        shakedown.uninstall_package_and_wait(SPARK_PACKAGE_NAME)
-        if _hdfs_enabled():
-            shakedown.uninstall_package_and_wait(HDFS_PACKAGE_NAME, HDFS_SERVICE_NAME)
-            _run_janitor(HDFS_SERVICE_NAME)
+    shakedown.uninstall_package_and_wait(SPARK_PACKAGE_NAME)
+    if utils.hdfs_enabled():
+        shakedown.uninstall_package_and_wait(HDFS_PACKAGE_NAME, HDFS_SERVICE_NAME)
+        _run_janitor(HDFS_SERVICE_NAME)
 
 
 @pytest.mark.sanity
 def test_jar():
-    master_url = ("https" if _is_strict() else "http") + "://leader.mesos:5050"
+    master_url = ("https" if utils.is_strict() else "http") + "://leader.mesos:5050"
     spark_job_runner_args = '{} dcos \\"*\\" spark:only 2 --auth-token={}'.format(
         master_url,
         shakedown.dcos_acs_token())
     jar_url = _upload_file(os.getenv('TEST_JAR_PATH'))
-    _run_tests(jar_url,
+    utils.run_tests(jar_url,
                spark_job_runner_args,
                "All tests passed",
                ["--class", 'com.typesafe.spark.test.mesos.framework.runners.SparkJobRunner'])
@@ -70,54 +57,12 @@ def test_jar():
 
 @pytest.mark.sanity
 def test_teragen():
-    if _hdfs_enabled():
-        _run_teragen()
-
-
-def _run_teragen():
-    jar_url = TERASORT_JAR
-    input_size = os.getenv('TERASORT_INPUT_SIZE', '1g')
-    _run_tests(jar_url,
-               "{} hdfs:///terasort_in".format(input_size),
-               "Number of records written",
-               ["--class", "com.github.ehiggs.spark.terasort.TeraGen",
-                "--conf", "spark.cores.max={}".format(TERASORT_MAX_EXECUTOR_CORES)])
-
-
-@pytest.mark.soak
-def test_terasort():
-    if _hdfs_enabled():
-        _delete_hdfs_terasort_files()
-        _run_teragen()
-        _run_terasort()
-        _run_teravalidate()
-
-
-def _run_terasort():
-    jar_url = TERASORT_JAR
-    _run_tests(jar_url,
-               "hdfs:///terasort_in hdfs:///terasort_out",
-               "",
-               ["--class", "com.github.ehiggs.spark.terasort.TeraSort",
-                "--conf", "spark.cores.max={}".format(TERASORT_MAX_EXECUTOR_CORES)])
-
-
-def _run_teravalidate():
-    jar_url = TERASORT_JAR
-    _run_tests(jar_url,
-               "hdfs:///terasort_out hdfs:///terasort_validate",
-               "partitions are properly sorted",
-               ["--class", "com.github.ehiggs.spark.terasort.TeraValidate",
-                "--conf", "spark.cores.max={}".format(TERASORT_MAX_EXECUTOR_CORES)])
-
-
-def _delete_hdfs_terasort_files():
-    job_name = 'hdfs-delete-terasort-files'
-    LOGGER.info("Deleting hdfs terasort files by running job {}".format(job_name))
-    dcosjob.add_job(job_name)
-    dcosjob.run_job(job_name, timeout_seconds=300)
-    dcosjob.remove_job(job_name)
-    LOGGER.info("Job {} completed".format(job_name))
+    if utils.hdfs_enabled():
+        jar_url = 'https://downloads.mesosphere.io/spark/examples/spark-terasort-1.0-jar-with-dependencies_2.11.jar'
+        utils.run_tests(jar_url,
+                   "1g hdfs:///terasort_in",
+                   "Number of records written",
+                   ["--class", "com.github.ehiggs.spark.terasort.TeraGen"])
 
 
 @pytest.mark.sanity
@@ -126,7 +71,7 @@ def test_python():
     python_script_url = _upload_file(python_script_path)
     py_file_path = os.path.join(THIS_DIR, 'jobs', 'python', 'PySparkTestInclude.py')
     py_file_url = _upload_file(py_file_path)
-    _run_tests(python_script_url,
+    utils.run_tests(python_script_url,
                "30",
                "Pi is roughly 3",
                ["--py-files", py_file_url])
@@ -145,7 +90,7 @@ def test_kerberos():
 
     principal = "nn/ip-10-0-2-134.us-west-2.compute.internal@LOCAL"
     keytab = "nn.ip-10-0-2-134.us-west-2.compute.internal.keytab"
-    _run_tests(
+    utils.run_tests(
         "http://infinity-artifacts.s3.amazonaws.com/spark/sparkjob-assembly-1.0.jar",
         "hdfs:///krb5.conf",
         "number of words in",
@@ -159,7 +104,7 @@ def test_kerberos():
 def test_r():
     r_script_path = os.path.join(THIS_DIR, 'jobs', 'R', 'dataframe.R')
     r_script_url = _upload_file(r_script_path)
-    _run_tests(r_script_url,
+    utils.run_tests(r_script_url,
                '',
                "Justin")
 
@@ -167,7 +112,7 @@ def test_r():
 @pytest.mark.sanity
 def test_cni():
     SPARK_EXAMPLES="http://downloads.mesosphere.com/spark/assets/spark-examples_2.11-2.0.1.jar"
-    _run_tests(SPARK_EXAMPLES,
+    utils.run_tests(SPARK_EXAMPLES,
                "",
                "Pi is roughly 3",
                ["--conf", "spark.mesos.network.name=dcos",
@@ -190,16 +135,12 @@ def test_s3():
             "spark.mesos.driverEnv.AWS_SECRET_ACCESS_KEY={}".format(
                 os.environ["AWS_SECRET_ACCESS_KEY"]),
             "--class", "S3Job"]
-    _run_tests(_upload_file(os.environ["SCALA_TEST_JAR_PATH"]),
+    utils.run_tests(_upload_file(os.environ["SCALA_TEST_JAR_PATH"]),
                app_args,
                "",
                args)
 
     assert len(list(s3.list("linecount-out"))) > 0
-
-
-def _hdfs_enabled():
-    return os.environ.get("HDFS_ENABLED") != "false"
 
 
 def _require_hdfs():
@@ -214,17 +155,6 @@ def _require_spark():
 
     _require_package(SPARK_PACKAGE_NAME, _get_spark_options())
     _wait_for_spark()
-
-
-def _require_spark_cli():
-    LOGGER.info("Ensuring Spark CLI is installed.")
-    installed_subcommands = dcos.package.installed_subcommands()
-    if any(sub.name == SPARK_PACKAGE_NAME for sub in installed_subcommands):
-        LOGGER.info("Spark CLI already installed.")
-    else:
-        LOGGER.info("Installing Spark CLI.")
-        shakedown.run_dcos_command('package install --cli {}'.format(
-            SPARK_PACKAGE_NAME))
 
 
 # This should be in shakedown (DCOS_OSS-679)
@@ -253,7 +183,7 @@ def _wait_for_spark():
 
 
 def _get_hdfs_options():
-    if _is_strict():
+    if utils.is_strict():
         options = {'service': {'principal': 'service-acct', 'secret_name': 'secret'}}
     else:
         options = {"service": {}}
@@ -273,14 +203,14 @@ def _is_hdfs_ready(expected_tasks = DEFAULT_HDFS_TASK_COUNT):
 
 
 def _get_spark_options():
-    if _hdfs_enabled():
+    if utils.hdfs_enabled():
         options = {"hdfs":
                    {"config-url":
                     "http://api.hdfs.marathon.l4lb.thisdcos.directory/v1/endpoints"}}
     else:
         options = {}
 
-    if _is_strict():
+    if utils.is_strict():
         options.update({'service':
                         {"principal": "service-acct"},
                         "security":
@@ -296,7 +226,7 @@ def _install_spark():
                {"config-url":
                 "http://api.hdfs.marathon.l4lb.thisdcos.directory/v1/endpoints"}}
 
-    if _is_strict():
+    if utils.is_strict():
         options['service'] = {"user": "nobody",
                               "principal": "service-acct"}
         options['security'] = {"mesos": {"authentication": {"secret_name": "secret"}}}
@@ -315,10 +245,6 @@ def _install_spark():
     shakedown.spinner.wait_for(pred)
 
 
-def _is_strict():
-    return os.environ.get('SECURITY') == 'strict'
-
-
 def _run_janitor(service_name):
     janitor_cmd = (
         'docker run mesosphere/janitor /janitor.py '
@@ -326,34 +252,6 @@ def _run_janitor(service_name):
     shakedown.run_command_on_master(janitor_cmd.format(
         svc=service_name,
         auth=shakedown.dcos_acs_token()))
-
-
-def _run_tests(app_url, app_args, expected_output, args=[]):
-    task_id = _submit_job(app_url, app_args, args)
-    LOGGER.info('Waiting for task id={} to complete'.format(task_id))
-    shakedown.wait_for_task_completion(task_id)
-    log = _task_log(task_id)
-    LOGGER.info("task log: {}".format(log))
-    assert expected_output in log
-
-
-def _submit_job(app_url, app_args, args=[]):
-    if _is_strict():
-        args += ["--conf", 'spark.mesos.driverEnv.MESOS_MODULES=file:///opt/mesosphere/etc/mesos-scheduler-modules/dcos_authenticatee_module.json']
-        args += ["--conf", 'spark.mesos.driverEnv.MESOS_AUTHENTICATEE=com_mesosphere_dcos_ClassicRPCAuthenticatee']
-        args += ["--conf", 'spark.mesos.principal=service-acct']
-    args_str = ' '.join(args + ["--conf", "spark.driver.memory=2g"])
-    submit_args = ' '.join([args_str, app_url, app_args])
-    cmd = 'dcos --log-level=DEBUG spark --verbose run --submit-args="{0}"'.format(submit_args)
-
-    LOGGER.info("Running {}".format(cmd))
-    stdout = subprocess.check_output(cmd, shell=True).decode('utf-8')
-
-    LOGGER.info("stdout: {}".format(stdout))
-
-    regex = r"Submission id: (\S+)"
-    match = re.search(regex, stdout)
-    return match.group(1)
 
 
 def _upload_file(file_path):
@@ -366,13 +264,3 @@ def _upload_file(file_path):
 
     basename = os.path.basename(file_path)
     return s3.http_url(basename)
-
-
-def _task_log(task_id):
-    cmd = "dcos task log --completed --lines=1000 {}".format(task_id)
-    LOGGER.info("Running {}".format(cmd))
-    stdout = subprocess.check_output(cmd, shell=True).decode('utf-8')
-    return stdout
-
-def _do_teardown():
-    return os.environ.get("DO_TEARDOWN") != "false"
